dcs/core_dc/DcSliceController/requirements.txt
mysqlclient==1.4.2
flask-restful==0.3.7
paramiko==2.4.2
cryptography==2.4.2
ruamel.yaml==0.15.90
pyyaml==5.1
libvirt-python==5.2.0
logger==1.4
ovs==2.10.0
ryu==4.31
xmltodict==0.12.0
